pyproject.toml
[tool.poetry]
name = "image_changer"
version = "0.1.0"
description = "A tool to modify images to avoid photo hashing detection"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^10.0.0"
click = "^8.1.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
